fix display of top-aligned buttons in ie11
This change should fix the issue with top-aligned buttons not displaying correctly in ie11.  Changing this flexbox setting didn't seem to have any effects on other browsers - but they were all displaying fine.

I'm far from an expert with CSS so I'm not certain that this is the best fix, but it is a fix.